A few more changes:
1. Consolidate use-wid and use-non-wid.

After chatting with ojarjur
I realized the agent must reside on a node pool that uses GCE_METADATA
server in order for the DL inverting proxy instances we use to authenticate
the VM.

The user pool is configured to always use a WID. No permission is
granted to the robot account, but the admin can grant access to cloud
API or resource accordingly down the road.

Therefore so far the only 'sane' configuration is to use WID for user
pool and non WID for default(hub+agent) pool.

2. Use a PVC to persist user data under /home/jovyan.
Be aware that gid 100 is hard coded in the config map. This need to
match the jupyter single user docker image.

3. Move the cloud service account setup to create infra.sh -- I think
all of these belong to the scope of terraform.

4. fix get_proxy_url.py for python3.8. The unicode cast looked very
strange to my eyes (I think even on windows the encoding conversion of
command args are automatic).

5. Install gcloud to the basic image.